Don't pass enable_snat by default in Router
The patch removes the default for enable_snat in the Router resource and
removes the value if not specified, to make it possible for non-admin
users to create routers: a policy rule in neutron forbids non-admin to
pass the enable_snat value.

It doesn't break backward compatibility as enable_snat is the default in
neutron.

Co-Authored-By: ala.rezmerita@cloudwatt.com
Closes-Bug: #1306593
Change-Id: Ib9d31f7e0a246bcaa663aaa74755526f8e31df7e
(cherry picked from commit d7836e922577f8fad4a2fba4658d4f74ed8a2e4a)