Implement resize of Cinder volumes
This patch adds update capabilities to volume size.
The new size must be bigger than old one (as supported by Cinder).
If the volume is attached, it is first detached, then resized and
attached again to the same mount on the same server.

Known limitations:
user has to care for herself about (un)mounting the volume
from within the instance OS and resizing the file system after
resizing the volume.

Change-Id: I7c54b18eb865d0786ec5e3f50c92aeabdaa80a64
Implements: blueprint update-cinder-volume