Make events return by insertion order
We fixed the table by having the ID available, but the query uses an
index, so MySQL will return the rows in index order, which is not
necessarily primary key order. We need to explicitly ask that it be
returned in primary key order. This should be a free operation on MySQL.

Change-Id: I48d6260e01736fc7475c13cf736968dbf8eda206