Eager-load Stack.raw_template for Stack.load
Whenever a stack is loaded from the database to perform a
parser.Stack.load, the raw_template should be prefetched to
avoid an extra sql query being performed.

However there are many calls to db_api.stack_get which do not need
to access Stack.raw_template, so the default behaviour is to
lazy-load it.

This change reduces the queries for a Stack.load from 3 to 2, so it
contributes to Related-Bug: #1306743

Change-Id: Ie6dad6ea4b1655a99e97a9ac2b7c00018406cd91