Pass unresolved template to Resource.update()
Previously the 'after' template passed to Resource.update() was completely
resolved. This led to an odd situation where the contents of a resource's
template snippet were either unresolved (but parsed) or resolved depending
on whether the resource had been updated or not. Resolve (ha!) this
discrepancy by passing a version of the template with the intrinsic
functions unresolved.

Change-Id: I8c11cdc3c28661db689b416833bde34cea12d98a