Restore resource_id from backup_stack when delete
When UpdateReplace is occured, the resource is moved to backup_stack,
and create new one. Then, resources which has dependency for updated
resource will be moved to backup_stack. If the creation of update is
failed, dependencies between backup_stack and exist_stack will exist.

Heat deletes backup_stack before deleting exist_stack. The dependency
still exists, so it will be failed. Thus, add behavior that restores
resource_id in the backup_stack when the stack is deleted.

Change-Id: Ib2a39fa20089b7140fc676f54db8cf66a4581c60
Closes-bug: #1270775