Change Volume metadata attribute to raw dict.
Currently the volume metadata attribute is json
serialized, but outputting the raw dict is more
convenient for human users. This is fixed by
introducing a new attribute 'metadata_values'
whose output is in raw dict form. The attribte
'metadata' is preserved for those users who are
already used to its json serialized output.

Change-Id: Ib95656fa8632449957cd8c24b232f4d1b09a0e42
Closes-Bug: #1322007