Use auth_token_info to initialize heat_keystoneclient sessions
Now we have the token info available in the request context, we can
avoid always re-requesting another token every time a client is
created.

This should be better for performance as it avoids one extra request to
keystone, and it also enables trust-scoped tokens to be more easily
used, since these cannot be used to request another token, thus
authentication will fail with the current scheme.

Change-Id: Id34a4104cedbd3f1c231bc0b9229a5cac7bb9402
Partial-Bug: #1317293