RPC API: Really don't include metadata in resource list
I made an attempt in e743158e7a61c00e2a32c340d3f268d0835f4c69 to not
include resource metadata when listing resources in the RPC API. Probably
due to a merge conflict, this was unsuccessful (although it did have the
effect of doubling the number of times we refreshed it from the database).

Neither the cfn API nor the native API passes on the metadata information
from list_stack_resources, so it is safe to remove this from the RPC API
for that call (which is the only one that passes detail=False) without
affecting the user-facing APIs.

Change-Id: Ibaddef9b290e88eb548c0427df5c007e3e14bc59