HOT intrinisic function list_join like Fn::Join
We have an increasing number of attributes which are a list of strings.

In HOT there is no way to join a list into a string as with the CFN
function Fn::Join.

This change implements HOT intrinsic function list_join using an
identical implementation as Fn::Join.

Tripleo will benefit from this function since the lack of a join
function in HOT is one of the barriers to transitioning to the
HOT template format.

Change-Id: I29e0b744271213429fff6922397b925eb9ccc204