Use stevedore to load client plugins
This uses stevedore to load any client plugin extension defined
in the heat.clients namespace in setup.cfg.

As a transition from non-plugin to plugin clients, requesting a
named client will first search for a client plugin, then fallback
to calling the _<name>() method.

Clients plugin initialise is triggered from resources initialise.

This change also adds the clients package for the constraints
plugin manager. Custom constraints often make client API
calls and are not necessarily associated with resources, so
in many cases the client plugin will be a better place for
constraints to live.

Change-Id: I853767944caeca2939c9494ae992a0c4f6474078