Add validation constraints on config inputs
This change adds a new property to the StructuredDeployment class to
enable a check on the config inputs. When this property is set to
'STRICT' we compare the inputs to the supplied input_values property and
we raise an exception if the input_values result in any inputs not being
assigned a value.
By default this is disabled.
I did a small refactor of the code adding a couple of extra methods.
Closes-bug: 1334283

Change-Id: If132163bcaf4eee61a9499808ab264e7b7f80cb2