Don't overwrite deployments metadata in Server resource
Currently, when metadata_update is called and local metadata has been
added for software deployments (e.g stuff not defined in the template)
we will silently discard the non-template-defined stuff when calling
metadata_update.

Arguably we are abusing the metadata section by pushing arbitrary additional
data into it but since that issue has already happened, as part of the
SoftwareDeployments implementation, the simplest fix seems to be to merge
the current and template re-resolved state of the metadata.

Change-Id: Id68d11b584734fbb828aabaca6cc75aab0f4ee4c
Closes-Bug: #1337772