Pass user_creds_id when creating backup stack
When creating a backup stack on update, we don't want to create a
new user_creds record, or update won't work with a trust-scoped
token.  So instead pass the existing stack user_creds_id in to the
constructor when creating the backup stack.

The current behavior is actually doubly wrong, since we don't delete
user_creds records when deleting the backup stack, so the stale
creds records are probably left behind in the DB.

Change-Id: I6bd0568fdf5301b876c40acddc73542bcfc224e3
Partial-Bug: #1342593