timeutils function retry_backoff_delay
Calculates an exponential backoff delay with jitter for retrying
operations which might fail.

The exponential delay on each attempt gives the service a chance
to process its backlog of retries after it starts responding again.

The jitter adds some randomness to the delay to prevent
multiple synchronised backoff retries from hitting the
service at exactly the same time.

The next change in this series makes use of this function.

Change-Id: Ie209ef62c0da68a5b503354764e5add2b6e5569a