Add language attribute to code samples
Create the literal_block nodes with a 'language' attribute to add syntax
highlighting in the generated documentation.

This change is not visible in the HTML output, but is used when the
generated XML is imported in the openstack-manuals hot-reference book
(WIP).

The 'hot' language is not currently supported in sphinx or in the
manuals, but it doesn't hurt to have it and it could be implemented
later.

Change-Id: Ie5d4ab31665da812ccd7ec1232d3858786170477
Implements: blueprint heat-templates
Closes-Bug: #1360559