Enable SoftwareDeployment for SoftwareComponent
This patch enables the existing SoftwareDeployment resource to cope
with the new SoftwareComponent resource, i.e. allow for associating a
SoftwareComponent resource to a server via a SoftwareDeployment.

In case of a SoftwareComponent resource, the SoftwareDeployment
resource will act at each resource action and create a derived
config to be passed to in-instances tools (instead of looking at its
own 'actions' property).

Note: to activate correct processing of SoftwareComponent data on
an instance, an additional os-collect-config plug-in similar to
55-heat-config will be implemented and posted to the heat-templates
repository. 55-heat-config will ignore configuration data with a
'group' property of 'component'.
The new in-instance hook will analyze component configurations,
derive the configuration to be applied based on the current resource
action and then invoke the appropriate configuration hook.

Implements: partial-blueprint action-aware-sw-config

Change-Id: I45d4d373babe0fff9b109cba97ce7ebcb90562df