Update stack.load_all adding show_nested
Add a show_nested argument, defaulted to false, which will
show nested stacks, not just top-level ones.  Due to an
unfortunate collision in the way backup and nested stacks
are currently stored (both are identifiable by having an
owner_id), we have to filter the backup stack by name, if
it exists.  In a future patch this can be cleaned up by
modifying the DB schema to store data to more explicitly
specify backup stacks.

blueprint: list-nested
Change-Id: I1b2e44380cf2f9162dbc4d9bc300e15c325f0eca