Add Server software_config_transport POLL_TEMP_URL
POLL_TEMP_URL will create a swift object with a newly
created UUID as the object name, in a container that matches
physical_resource_name.

Two TempURLs are created to this object, one to GET object
contents, and another to PUT.

The GET TempURL is used for the os-collect-config configured
metadata_url. The PUT url is stored in resource data so that
it can be retrieved in EngineService when deployment data
needs to be pushed out.

The object name is also stored in resource data so that the
object can be deleted during handle_delete. The container
is also deleted if deleting the object results in an empty
container.

Partial-Blueprint: swift-deployment-transport
Change-Id: Id644882a110171c12db8c34cfde02afc3d1d8626