Don't stop creates/updates immediately on error
When an error occurs during a stack create or update operation, don't
immediately stop the operation. Instead, wait up to 4 minutes for resoures
already in progress to complete before forcibly cancelling them. This
minimises the number of resources that will need to be replaced when using
the features from the blueprint update-failure-recovery to recover after
such a failed operation.

Change-Id: Ib08f3d8faea31f251a46fc0537983127f8ba2783