Split scaling policy into separate files
A further step to split the AWS and the OS version of auto-scaling group
related classes.  This patch splits the scaling policy and launch
configuration into separate files.

This patch is not breaking any existing test cases. But more test cases
will be needed for the OS version scaling policy.

Implements: partial-blueprint reorg-asg-code
Change-Id: I0f888496d92b0efb5d63b567864488ff8a45e4ac