Clean up signalling events after stack updates
We provide an Event object to every stack update in order to signal it that
we need to cancel the update; this change ensures that we delete it when
the thread is complete. Previously we were leaking memory on every update.

Change-Id: Ie0fbe3ec3ab57a2cb9dd5e551db15285b2b423c0
Closes-Bug: #1376857
(cherry picked from commit 6f13c2d0a4ff2004654a690c1e84a0c9d40ed577)