Don't update a LoadBalancer under autoscaling control
Unfortunately, Autoscaling currently uses the update() method of a
LoadBalancer resource to do updates, with the result that the current
member list gets persisted (good) and used to compare in the event of
future *stack* updates (bad).

With this patch, we assume that LoadBalancers under the control of
Autoscaling will never have a members list property supplied in the
template.  We then ignore any updates to Autoscaling LoadBalancers that
don't actually modify the template.

The test changes revert the changes made in order to be able to merge
d32370233eaf2a5c32888f269bd1dc5e0e787467, before which LoadBalancers were
behaving correctly.

Change-Id: I9c02ab3d3dfbee0a8a90dd0ba345a5acdaf8a610
Closes-Bug: #1379619