Add tox genconfig target
It's easy to get a false config generated using a local (non venv)
environment to run generate_sample.sh, so this makes it easier to
run the config generator in a venv for consistent results which
should match what's done in the gate.

Possibly in future we could consider removing the static conf.sample
like some other projects have already done and just documenting
this as the method to generate the sample.

Change-Id: I92964d75e0adf73eb0935c2052adf344f15bfe95