Ensure all Neutron LoadBalancer members are deleted
Previously we only deleted the ones listed in the template, even though
members can be added by autoscaling that don't appear in the template. In
theory this is OK if we're deleting the autoscaling group too, and since
it's not possible to update the LoadBalancer property on an autoscaling
group I can't find a way to actually trigger any bug (hence no new test).
However, it's much safer to ensure we delete all of the remaining members
when we delete the LoadBalancer resource.

Change-Id: I6f1c02ae408e9169a23ae32e7fce5ea0d2239add
Closes-Bug: #1310602
(cherry picked from commit 3f33130b89b96c6fcc1dace12498f691413e4713)