Halt test_cinder_volume_create_backup_restore on error
test_cinder_volume_create_backup_restore fails frequently due to issues
outside heat. Instead of failing, this change halts the test when the
error condition is triggered. This will allow some of the test to run
while making the check job voting.

Once logging is configured in heat_integrationtests an elastic-recheck
search can monitor for the frequency of this error to aid
nova/cinder/swift developers to fix it.

Change-Id: I09722ad725a8d23fb2028c17b0dd9fcab3957649
Related-Bug: #1382300