Failed res no need UpdateReplace which has nested_stack
Now update the resources in FAILED status will enter UpdateReplace
flow, but it's not fit for resources which has nested stack.
Especially for autoscaling group, if there is one or more server
were created/deleted failed, we shouldn't destroy all the servers
and then create new servers for the group while updation.

Change-Id: Ib53f66fe481acc282123dc1743ef8310456a1d69
Closes-bug: #1383618