Repair tox -ecover
Some tests are ignored during the invocation of 'tox -epy27' and
similar, but they were not ignored during the 'cover' profile. Becuase
the ignored tests actually fail when run this means that 'tox -ecover'
could not be used.

This patch just ignores the same tests in the 'cover' profile as the
default.

Change-Id: I7952c312addfed87f6055c1a360adfbf1fd5b168