RandomString physical_resource_id as id not the string
If someone choses a string length of over 255 we overflow the DB
id field causing an error, so instead use the physical name 
for the resource_id.

The actual string is already stored (encrypted) in the resource_data
so it's probably better anyway to refer to the value in one place
and just override FnGetRefId to maintain the existing interface behavior.

Change-Id: Ide8899a2aa0a9fed037835dc46225f52c291ace6
Closes-Bug: #1392359