ResourceGroup don't special-case zero count
Currently, when you specify an initial count of zero, we don't
create the nested stack.  Unfortunately this then means you can
never update it, which makes the group forever broken..

Instead, always create the stack on create, even if it's an empty
stack containing zero resources.  This allows subsequent updates
which move to a non-zero count to work as expected.

Change-Id: I494229fcf85a1c28428fa2925c060fc1f40f4f36
Closes-Bug: #1392796