SoftwareDeployments make servers property updateable
Allow the servers list to be updated, which will cause an update
of the resource group instead of replacing it.

This looks inconvenient to test in the unit tests due to needing
to create a coupled nested stack (which will further impede the
decouple-nested work), so I'll post a followup patch with a
functional test instead, if that's viewed as acceptable.

Change-Id: I1714d2cef2d0edbeb82c16e5a83acc458009bd47
Closes-Bug: #1389178