Don't block stack deletion if user creds fail
In some deployments, user passwords are generated by other
authentication infrastructures where the length of the Base64 encoded
credentials is longer than the DB limit.  When this incorrect behavior
happens silently, the user credential can not be retrieved.
In this case, stack deletion will abort due to uncaught exception.

This patch enables a stack deletion to handle this error so that stack
can be properly deleted.

Change-Id: Iaf80b18a83c7752465621c232335dcf32afbcbf3
Closes-Bug: 1386213