Deprecate HARestarter resource type
This was basically a failed experiment in High Availability. Amongst other
things - for example it is perpetually causing users to misunderstand how
it works - it relies on the paradigm of a resource just being a place to
stick code that can then go and perform operations on the stack itself
which are not defined in Heat's lifecycle model, rather than purely as a
representation of some real OpenStack resource. We later realised that this
was the cause of many of our worst headaches. Convergence will be a far
superior solution to the problems that HARestart was designed to solve.

HARestarter is an architectural dead-end, and we need to communicate that
much more clearly to users along what it actually can and cannot do.

Change-Id: I4ad9c7ae54e3f04224b780ecb24d0827951d18da