Prevent showing 'show' too early
The logic in format_resource_attributes was succeptible to a failure
when resource that has 'show' attribute and is early IN_PROGRESS has
that attribute not populated yet, returning an empty string instead.

That might happen when asking for resource-show soon enough after
stack-create, and that is what test_server_cfn_init in
heat_integrationtests effectively does.

Change-Id: I37e209042048cb4409fcd27fcf9587b594f3129b
Closes-Bug: #1401107