Wait for Sahara Cluster to be deleted
Currently the stack with cluster resource depending on cluster template
can not be deleted as the deletion of the cluster template fails since
actually the dependent cluster is not completely deleted yet.

This patch implements check_delete_complete for OS::Sahara::Cluster resource,
allowing the resource to wait until the actual Sahara cluster is deleted
before proceeding along dependency graph.

Change-Id: I9bf3192b34081f045f8165a3b958a7f9a6418ff7
Closes-Bug: #1400540