Move LB reload logic into scaling library
The lb_reload() logic is moved into a separate module of the scaling
library.  A new test case is added to test this utility function.

Since the revised _lb_reload() method checks if LOAD_BALANCER_NAMES is
defined/provided or not, the OpenStack version of AutoScalingGroup
doesn't have to implement an empty _lb_reload() method now.

partial-blueprint: as-lib
partial-blueprint: reorg-asg-code

Change-Id: I7387d0f27c6121be29d9a035be4cca0b9bbdf441