Better error on deleting non-empty Swift container
Swift client method delete_container() does not allow to delete
non-empty containers, in which case user gets a confusing 409 error
exessively truncated by Swift with no actual reason information.

This patch catches 409 errors from Swift client on container delete,
checks for objects in the container and raises a suitable error if
container is not empty, so that the DELETE_FAILED reason is clearly
seen.

Change-Id: Ibca3851d11b509413b739a693b1ddd244479d37b
Closes-Bug: #1406263