Add support status for network/subnet neutron properties
During Juno, support was added to various neutron resources for
lookup of network/subnet by both name and ID, deprecating the
previous *_id properties in the process.  Some users have been
confused by when the new properties became available, so add
the SupportStatus to the schema so the docs clarify it.

Change-Id: I0631f95d9057afadedc32dc7a94b02547e522576
Related-Bug: #1286128