Make the update policy timeout check into a unit test
There are a lot of repeated tests here, the one that looks unique
is the one that tests a timed out replacement.
This is an instance group method so it is moved to test_instance_group and
it is reworked to not need a stack creation.

Part of blueprint decouple-nested
Change-Id: I4406ee772568ea24edfb34e43891f600e9de4573