Fix the standalone authentication mechanism
The keystoneclient used by the password_auth may be a v2 or v3 client.
This means that we shouldn't access arbitrary attributes from the client
as they many not exist on both.

We really don't care about the version of keystoneclient in this case
all we want to do is create a session for getting a token, so copy the
session creation logic out to auth_password.

Change-Id: I1b1f3e3f5822c0bba2dc2e8f592760dd5dba7fbc
Closes-Bug: #1416170