Pass the resource name into get_resource_info()
Problem
If you have the following in the environment, the resource "a_res" will
be created with "other.yaml" instead of "my.yaml".

resource_registry:
  resources:
    a_res:
      OS::Nova::Server: my.yaml
  OS::Nova::Server: other.yaml

The reason for this is that the resource name has not been passed into
get_resource_info().

Note: a functional test is coming in a later commit.

Closes-bug: 1418935
Change-Id: I45a453ac342eab287544fc9d4f94e9bd5716bb8e