Always check for 'component' in SoftwareDeployment
When SoftwareDeployment is associated to a SoftwareComponent, config
hooks in the instance should always be trigger and the in-instance
logic will decide whether or not anything has to be done.

The check whether the SoftwareDeployment resource is associated to a
SoftwareComponent resource is implemented in _handle_action which is
called by all handle_ methods.
For handle_delete, however, the 'actions' property of
SoftwareDeployment was evaluated first and only if it contained
DELETE the _handle_action method was invoked.
This patch make handle_delete also directly call _handle_action to
check for presence of a SoftwareComponent like all other lifecycle
hooks do.

Change-Id: Ibb813db23fa7c9ab20435887f6cc49deaaf51de2
Closes-Bug: #1416994