Catch Conflict exc for floating ip association
When heat is trying to delete floating ip association and nova
can't find an association between floating ip and server
nova raises conflict exception instead of not_found. It happens
because nova assumes that request was incorrect (floating ip
is existing but link is not actual).
So heat need to support such kind of cases.

Change-Id: I3cc1fb13b9d3f0c1a2f7ee1c8ae7a99160b28951
Closes-bug: #1418878